r1802/sofa-dev : ADD : "ENDING_EVENT" is a new type of event to notify the end of a sequence of topological change events => it can be use later for synchronisation issues.
"notifyEndingEvent()" is a new high-level method which is called after the handling of a whole topological modification (such as triangular incision or tetrahedron removal) to add the ENDING_EVENT in the TopologicalChange list. Then the method "propagateTopologicalChanges()" is called to propagate all the events of the list.

MODIFY : a good stable version of scene "cuttingMembraneUnderGravity.scn" and some small other modifications.

 //partial sync (5/7 changes)//


git-svn-id: svn://scm.gforge.inria.fr/svnroot/sofa/dev/trunk/Sofa@1316 b7f1a160-662a-0410-a6d8-bea165342f16