r3020/sofa-dev : NEW: compatibility with mapped Mass and dynamics i.e. The computation of the inertial force created on mapped object due to the velocity of the input object. (coriolis type forces) map->propagateA() is added in propagatePositionAndVelocity visitor.
It calls 'computeAccFromMapping' is a mapping that computes the acceleration of 'toModel' due to the velocity of 'fromModel'

(this function only implemented for RigidRigidMapping... but has to be done for RigidMapping and ArticulatedSystemMapping)

This acceleration multiplied by the mass in UniformMass (has to be done for other mass) during addForce function

In testMappedMassDynamics a simple example: a rigidmass is mapped by a rigidrigidMapping. The computation is made where the transparent red box is placed. 

 You can desactivate "compute_mapping_inertia"  in UniformMass to see the difference 

 //partial sync (10/14 changes)//


git-svn-id: svn://scm.gforge.inria.fr/svnroot/sofa/dev/trunk/Sofa@2249 b7f1a160-662a-0410-a6d8-bea165342f16