r9826/sofa-dev : Added flag «symmetricMatrix()» to the MechanichalParams. This is used to state that the matrix of the equation system is supposed symmetric. If so, the conjugate gradient (or other solvers specialized on symmetric matrices) can be safely applied. The default value is: true.
Currently, this parameter is used to skip the calls to Mapping::applyDJT in MechanicalAddMBKdxVisitor and MechanicalComputeDfVisitor, because applyDJT corresponds to a non-symmetric matrix which can mess up the Conjugate Gradient solution.

So this method is never used, and the stiffness is not exactly computed when nonlinear mappings are applied.

This is a motivation, and a testbed, for future nonsymmetric solvers.

 


git-svn-id: svn://scm.gforge.inria.fr/svnroot/sofa/dev/trunk/Sofa@6598 b7f1a160-662a-0410-a6d8-bea165342f16