r11868/sofa-dev : Ongoing work on switching between master models based on distance from the camera.
Warning: added standard opengl matrices to the VisualParams. There may be some redundancy with member sceneTransform, which is not documented.

Added method BaseMapping::computeLocalCoordinates() to force computation of the local coordinates based on the slave coordinates. Currently, only RigidMapping uses it, the other mappings still do it directly in their init() method.


git-svn-id: svn://scm.gforge.inria.fr/svnroot/sofa/dev/trunk/Sofa@7952 b7f1a160-662a-0410-a6d8-bea165342f16