r12240/sofa-dev : Added a command line option in runSofa to switch on the display of computation time, and made the hierarchy more visible in the output.
Created Compliant/examples/CantileverBeam_4_4_20.scn to compare with the assembly and solution used in examples/Components/linearsolver/CSPARSE-SparseCholeskySolver.scn. The new assembly is slower, in the case of FEM without mappings. The result of the simulation is different also, the Flexible FEM have a non-symmetric behavior compared with the classical FEM.

 //partial sync (11/13 changes)//


git-svn-id: svn://scm.gforge.inria.fr/svnroot/sofa/dev/trunk/Sofa@8256 b7f1a160-662a-0410-a6d8-bea165342f16