push bce568c..1962b53
* commit 1962b53880239977f15c56ffba75254242e21801
cmake: Redo the remove_duplicate from 10889; doesn't cause a weird bug.

Still, this reveals a bug: the first time I run cmake after enabling
SOFA-MISC_TESTS, the include directories for gtest are somehow
forgotten, and the compilation fails. The problem disappears if I run
cmake a second time. It was hidden until now because cmake choked on an
empty list when trying to REMOVE_DUPLICATE, and thus was always run a
second time.

   1 files changed, 1 insertions(+), 1 deletions(-)



git-svn-id: svn://scm.gforge.inria.fr/svnroot/sofa/trunk@10894 b7f1a160-662a-0410-a6d8-bea165342f16