push f8e626b..892f90b
* commit 892f90b2f3e6760a7afa5ad4dbff6b4420740fe5
Modeler: replace some Qt3Support widgets with Qt4 ones

This commit is a first step toward pure Qt4 applications; it also does
quite a bit of cleanup, which is long overdue. Removing Q3Support
widgets is the next step after getting rid of Qt3 code: Qt3Support is
part of Qt4, but not is not supported by Qt5. Pure Qt4 is surprisingly
cleaner and more consistent. Besides, it looks and feels a lot nicer :-)

UI files are dropped in favor of plain Qt-based C++ code: although they
seem convenient at first sight, they are actually a pain to maintain and
evolve.

This is a work in progress. There should be no significant difference in
behavior, but feel free to complain it anything is broken.

   29 files changed, 395 insertions(+), 2106 deletions(-)



git-svn-id: svn://scm.gforge.inria.fr/svnroot/sofa/trunk@11228 b7f1a160-662a-0410-a6d8-bea165342f16